+++ packages/driver/cypress/e2e/issues/28527.cy.ts
@@ -1,0 +1,26 @@
+describe('issue 28527', { testIsolation: false, retries: 2 }, () => {
+  before(() => {
+    cy.visit('/fixtures/empty.html')
+  })
+
+  beforeEach(() => {
+    // ensure we run the tests in run mode
+    Cypress.config('isInteractive', false)
+    Cypress.config('exit', true)
+  })
+
+  // there can only be one test in this file to ensure we are testing the scenario
+  // where a test fails and the runner does not navigate to about:blank between retries
+  it('test 1', () => {
+    cy.then(() => {
+      // fail the first attempt to ensure we don't go to about:blank before the second attempt
+      if (Cypress.currentRetry < 2) {
+        throw new Error(`attempt ${Cypress.currentRetry + 1} error`)
+      }
+    })
+
+    cy.get('title').should('have.text', 'Empty HTML Fixture')
+    cy.url().should('include', '/fixtures/empty.html')
+    cy.url().should('not.include', 'about:blank')
+  })
+})